+++ one/tests/alf/test_alf_path.py
@@ -1,4 +1,5 @@
 """Unit tests for the one.alf.path module."""
+import os
 import unittest
 import tempfile
 from datetime import datetime
@@ -326,6 +327,45 @@ class TestALFPath(unittest.TestCase):
         alfpath = self.alfpath.relative_to_session()
         self.assertRaises(ALFInvalid, alfpath.with_revision, 'bar')
 
+    def test_with_collection(self):
+        """Test for PureALFPath.with_collection method."""
+        # Test with dataset alone
+        alfpath = self.alfpath.parents[2] / self.alfpath.name
+        expected = self.alfpath.parents[2] / 'collection' / self.alfpath.name
+        self.assertEqual(expected, alfpath.with_collection('collection'))
+        # Test with collection alone
+        alfpath = self.alfpath.parents[2] / 'alf' / 'foo'
+        expected = self.alfpath.parents[2] / 'collection'
+        self.assertEqual(expected, alfpath.with_collection('collection'))
+        # Test with revision alone
+        alfpath = self.alfpath.parents[2] / '#2020-01-01#'
+        expected = self.alfpath.parents[2] / 'collection' / '#2020-01-01#'
+        self.assertEqual(expected, alfpath.with_collection('collection'))
+        # Test with collection, revision, and dataset
+        expected = self.alfpath.parents[2].joinpath('collection', *self.alfpath.parts[-2:])
+        self.assertEqual(expected, self.alfpath.with_collection('collection'))
+        # Test with session folder
+        expected = self.alfpath.parents[2] / 'collection'
+        self.assertEqual(expected, self.alfpath.parents[2].with_collection('collection'))
+        # Test with backslashes (Windows only)
+        collection = r'\collection\foo_00'
+        expected /= 'foo_00'
+        if os.sep == '\\':
+            self.assertEqual(expected, self.alfpath.parents[2].with_collection(collection))
+        else:
+            # On non-Windows systems, backslashes are valid dir names but invalid ALF
+            self.assertRaises(ValueError, self.alfpath.parents[2].with_collection, collection)
+        # Test with trailing slashes
+        self.assertEqual(expected, self.alfpath.parents[2].with_collection('/collection/foo_00/'))
+        # Test with invalid collection name
+        self.assertRaises(ValueError, self.alfpath.with_collection, '#invalid#')
+        # Test with invalid path
+        alfpath = self.alfpath.parents[2].joinpath('#revision#', 'alf')
+        self.assertRaises(ALFInvalid, alfpath.with_collection, 'collection')
+        # Test without session parts
+        test_method = ALFPath(*self.alfpath.parts[-3:]).with_collection
+        self.assertRaises(ValueError, test_method, 'collection')
+
     def test_with_padded_sequence(self):
         """Test for PureALFPath.with_padded_sequence method."""
         # Test already padded
@@ -595,6 +635,12 @@ class TestALFPath(unittest.TestCase):
         # Number
         self.assertEqual('001', self.alfpath.sequence)
         self.assertEqual('', self.alfpath.relative_to_session().sequence)
+        # Collection
+        self.assertEqual('alf', self.alfpath.collection)
+        self.assertEqual('', self.alfpath.relative_to_session().collection)
+        # Revision
+        self.assertEqual('2020-01-01', self.alfpath.revision)
+        self.assertEqual('', self.alfpath.relative_to_session().revision)
         # session_parts
         self.assertEqual(('labname', 'subject', '1900-01-01', '001'), self.alfpath.session_parts)
         alfpath = ALFPath(*self.alfpath.parts[5:])